# docs/swagger/swagger.yaml
basePath: /api
definitions:
...
  handlers.BatchPermissionCheckRequest:
    properties:
      checks:
        items:
          $ref: '#/definitions/handlers.ResourceActionCheck'
        minItems: 1
        type: array
      user_id:
        type: string
    required:
    - checks
    - user_id
    type: object
  handlers.BatchPermissionCheckResponse:
    properties:
      results:
        additionalProperties:
          type: boolean
        type: object
    type: object
...
  handlers.PermissionCheckRequest:
    properties:
      action_slug:
        type: string
      resource_slug:
        type: string
      user_id:
        type: string
    required:
    - action_slug
    - resource_slug
    - user_id
    type: object
  handlers.PermissionCheckResponse:
    properties:
      allowed:
        type: boolean
      reason:
        type: string
    type: object
...
  handlers.ResourceActionCheck:
    properties:
      action_slug:
        type: string
      resource_slug:
        type: string
    required:
    - action_slug
    - resource_slug
    type: object
...
host: localhost:8000
info:
  contact:
    email: support@forgecrud.com
    name: API Support
    url: http://www.forgecrud.com/support
  description: Central API documentation - For all ForgeCRUD microservices
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: ForgeCRUD API
  version: "1.0"
paths:
...
  /permissions/batch-check:
    post:
      consumes:
      - application/json
      description: Check multiple resource-action permissions for a user in a single
        request
      parameters:
      - description: Batch permission check request
        in: body
        name: batch
        required: true
        schema:
          $ref: '#/definitions/handlers.BatchPermissionCheckRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Batch permission check results
          schema:
            $ref: '#/definitions/handlers.BatchPermissionCheckResponse'
        "400":
          description: Invalid request format
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Check multiple permissions
      tags:
      - permission-checks
...
  /permissions/check:
    post:
      consumes:
      - application/json
      description: Check if a user has permission for a specific resource and action
      parameters:
      - description: Permission check request
        in: body
        name: check
        required: true
        schema:
          $ref: '#/definitions/handlers.PermissionCheckRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Permission check result
          schema:
            $ref: '#/definitions/handlers.PermissionCheckResponse'
        "400":
          description: Invalid request format
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Check single permission
      tags:
      - permission-checks
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and the JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...